There can now be more than one RSS feed with the same link (it just has to be on different server/channel combination). RSS feeds now cache results, so if there are two feeds with the same link, it won't fetch the same result set twice. Fixed a compilation error in the bot I introduced via the web interface.
git-svn-id: https://vivio.googlecode.com/svn/trunk@73 2da491a5-97c4-c799-e8a7-435aa7d697a3

+++ Vivio/src/commands/ReminderCommand.java
@@ -6,7 +6,6 @@
  */
 package commands;
 
-import java.math.BigDecimal;
 import java.sql.SQLException;
 import java.sql.Timestamp;
 import java.util.HashMap;
@@ -151,7 +150,7 @@ public class ReminderCommand extends Command {
 							}
 							lastParsed = LAST_PARSED_DIGIT;
 							
-							currentNumber = currentNumber*10 + ((int)params.charAt(i) - '0');
+							currentNumber = currentNumber*10 + (params.charAt(i) - '0');
 						}
 						
 						else if(Character.isLetter(params.charAt(i))) {
@@ -182,7 +181,7 @@ public class ReminderCommand extends Command {
 						}
 					}
 					
-					moment += (long)currentNumber * getUnitToMillisecMultiplier(unit);
+					moment += currentNumber * getUnitToMillisecMultiplier(unit);
 					
 					break;
